requirements.txt
networkx>=3.1
gymnasium>=0.29.0
numpy>=1.24.0
matplotlib>=3.7.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: for Azure OpenAI powered user hints
openai>=1.30.0
weave>=0.50.0
python-dotenv>=1.0.0

# Web backend & API testing
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.24.1

# AgentBeats integration
mcp>=1.0.0
sse-starlette>=1.6.0
a2a-sdk>=0.3.0
typer>=0.12.3
rich>=13.0.0
